abstraction for DataClass and code improvements
